fix(daemon and shell)
